// src/lib/generated/types.gen.ts
export type GuideUpdateRequest = {
    nickname?: string;
    profileImageUrl?: string;
    location?: 'SEOUL' | 'BUSAN' | 'DAEGU' | 'INCHEON' | 'GWANGJU' | 'DAEJEON' | 'ULSAN' | 'SEJONG' | 'GYEONGGI' | 'GANGWON' | 'CHUNGCHEONGBUK' | 'CHUNGCHEONGNAM' | 'JEOLLABUK' | 'JEOLLANAM' | 'GYEONGSANGBUK' | 'GYEONGSANGNAM' | 'JEJU';
    description?: string;
};

export type ApiResponseGuideResponse = {
    msg: string;
    data?: GuideResponse;
};

export type GuideResponse = {
    id: number;
    email: string;
    nickname: string;
    profileImageUrl?: string;
    role: 'USER' | 'ADMIN' | 'GUIDE' | 'PENDING';
    location?: 'SEOUL' | 'BUSAN' | 'DAEGU' | 'INCHEON' | 'GWANGJU' | 'DAEJEON' | 'ULSAN' | 'SEJONG' | 'GYEONGGI' | 'GANGWON' | 'CHUNGCHEONGBUK' | 'CHUNGCHEONGNAM' | 'JEOLLABUK' | 'JEOLLANAM' | 'GYEONGSANGBUK' | 'GYEONGSANGNAM' | 'JEJU';
    description?: string;
};